Integrate CORS functionality directly into pipeline
